Commit for the fourth step of the assignment 2:
4.	Data display (3)
a.	On the home page, provide the following features
i.	Search for a keyword (it can be a part of title, desc, or even category) (15pts)
ii.	Filter based on the given category (all categories are to be provided a sa select box and choosing any category from the list filters the list of all the products) (15 pts)
b.	Use API documentation to explore available ones and their usage.